build/ci: Update Go to v1.24, golangci-lint to v1.64.5 and fix issues

+++ backtester/engine/grpcserver_test.go
@@ -90,7 +90,7 @@ func TestExecuteStrategyFromFile(t *testing.T) {
 		StrategyFilePath:  dcaConfigPath,
 		StartTimeOverride: timestamppb.New(time.Now().Add(-time.Hour * 6).Truncate(time.Hour)),
 		EndTimeOverride:   timestamppb.New(time.Now().Add(-time.Hour * 2).Truncate(time.Hour)),
-		IntervalOverride:  uint64(time.Hour.Nanoseconds()),
+		IntervalOverride:  time.Hour.Nanoseconds(),
 	})
 	if !errors.Is(err, nil) {
 		t.Errorf("received '%v' expecting '%v'", err, nil)
@@ -218,7 +218,7 @@ func TestExecuteStrategyFromConfig(t *testing.T) {
 	}
 
 	dataSettings := &btrpc.DataSettings{
-		Interval: uint64(defaultConfig.DataSettings.Interval.Duration().Nanoseconds()),
+		Interval: defaultConfig.DataSettings.Interval.Duration().Nanoseconds(),
 		Datatype: defaultConfig.DataSettings.DataType,
 	}
 	if defaultConfig.DataSettings.APIData != nil {
@@ -256,7 +256,7 @@ func TestExecuteStrategyFromConfig(t *testing.T) {
 	if defaultConfig.DataSettings.DatabaseData != nil {
 		dbConnectionDetails := &btrpc.DatabaseConnectionDetails{
 			Host:     defaultConfig.DataSettings.DatabaseData.Config.Host,
-			Port:     uint32(defaultConfig.DataSettings.DatabaseData.Config.Port),
+			Port:     defaultConfig.DataSettings.DatabaseData.Config.Port,
 			Password: defaultConfig.DataSettings.DatabaseData.Config.Password,
 			Database: defaultConfig.DataSettings.DatabaseData.Config.Database,
 			SslMode:  defaultConfig.DataSettings.DatabaseData.Config.SSLMode,
